# config/zedrem-CoffiDa-1.0.0-swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/zedrem/CoffiDa/1.0.0"
  } ],
  "info" : {
    "description" : "CoffiDa is platform for reviewing the best local coffee spots. Anybody can use CoffiDa for finding the best local Café’s, but Users who sign up for an account can also publish reviews. Reviews consist of a series of ratings (e.g. quality, price, cleanliness) along with a short body of text for describing their experience. Users can also comment on, and ‘like’ other reviews.",
    "version" : "1.0.0",
    "title" : "CoffiDa API",
    "contact" : {
      "name" : "Ashley Williams",
      "email" : "ashley.williams@mmu.ac.uk"
    }
  },
...
  "paths" : {
...
    "/location/{loc_id}/review" : {
      "post" : {
        "tags" : [ "Location Reviews" ],
        "summary" : "Add a new review",
        "operationId" : "addReview",
        "description" : "Will create a new review in the system\n",
        "security" : [ {
          "User" : [ ]
        } ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/loc_id"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddReview"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorised"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      }
    },
    "/location/{loc_id}/review/{rev_id}" : {
      "patch" : {
        "tags" : [ "Location Reviews" ],
        "summary" : "Update a review",
        "operationId" : "updateReview",
        "description" : "Update an existing review\n",
        "security" : [ {
          "User" : [ ]
        } ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/loc_id"
        }, {
          "$ref" : "#/components/parameters/rev_id"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateReview"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorised"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      },
      "delete" : {
        "tags" : [ "Location Reviews" ],
        "summary" : "Delete a review",
        "operationId" : "deleteReview",
        "description" : "Delete an existing review\n",
        "security" : [ {
          "User" : [ ]
        } ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/loc_id"
        }, {
          "$ref" : "#/components/parameters/rev_id"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorised"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      }
    },
...
    "/location/{loc_id}/review/{rev_id}/like" : {
      "post" : {
        "tags" : [ "Location Reviews" ],
        "summary" : "Like a review",
        "operationId" : "addReviewLike",
        "description" : "Add a like to an existing review\n",
        "security" : [ {
          "User" : [ ]
        } ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/loc_id"
        }, {
          "$ref" : "#/components/parameters/rev_id"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorised"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      },
      "delete" : {
        "tags" : [ "Location Reviews" ],
        "summary" : "Remove a like from a review",
        "operationId" : "deleteReviewLike",
        "description" : "Remove a like from an existing review\n",
        "security" : [ {
          "User" : [ ]
        } ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/loc_id"
        }, {
          "$ref" : "#/components/parameters/rev_id"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "401" : {
            "description" : "Unauthorised"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      }
    },
...
  },
  "components" : {
    "parameters" : {
...
      "loc_id" : {
        "name" : "loc_id",
        "in" : "path",
        "description" : "A unique integer ID for identifying a Location",
        "required" : true,
        "schema" : {
          "type" : "integer"
        }
      },
      "rev_id" : {
        "name" : "rev_id",
        "in" : "path",
        "description" : "A unique integer ID for identifying a Review",
        "required" : true,
        "schema" : {
          "type" : "integer"
        }
      },
...
    },
    "schemas" : {
...
      "overall_rating" : {
        "type" : "integer",
        "example" : 4,
        "minimum" : 0,
        "maximum" : 5
      },
      "price_rating" : {
        "type" : "integer",
        "example" : 2,
        "minimum" : 0,
        "maximum" : 5
      },
      "quality_rating" : {
        "type" : "integer",
        "example" : 3,
        "minimum" : 0,
        "maximum" : 5
      },
      "clenliness_rating" : {
        "type" : "integer",
        "example" : 5,
        "minimum" : 0,
        "maximum" : 5
      },
      "review_body" : {
        "type" : "string",
        "example" : "Great coffee, but the bathrooms stank!"
      },
...
      "AddReview" : {
        "type" : "object",
        "required" : [ "overall_rating", "price_rating", "quality_rating", "clenliness_rating", "review_body" ],
        "properties" : {
          "overall_rating" : {
            "$ref" : "#/components/schemas/overall_rating"
          },
          "price_rating" : {
            "$ref" : "#/components/schemas/price_rating"
          },
          "quality_rating" : {
            "$ref" : "#/components/schemas/quality_rating"
          },
          "clenliness_rating" : {
            "$ref" : "#/components/schemas/clenliness_rating"
          },
          "review_body" : {
            "$ref" : "#/components/schemas/review_body"
          }
        }
      },
      "UpdateReview" : {
        "type" : "object",
        "properties" : {
          "overall_rating" : {
            "$ref" : "#/components/schemas/overall_rating"
          },
          "price_rating" : {
            "$ref" : "#/components/schemas/price_rating"
          },
          "quality_rating" : {
            "$ref" : "#/components/schemas/quality_rating"
          },
          "clenliness_rating" : {
            "$ref" : "#/components/schemas/clenliness_rating"
          },
          "review_body" : {
            "$ref" : "#/components/schemas/review_body"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "User" : {
        "type" : "apiKey",
        "name" : "X-Authorization",
        "in" : "header"
      }
    }
  }
}